+++ internal/index/testdata/rails/app/models/article.rb
@@ -2,6 +2,8 @@ module Publishable
 end
 
 class Article
+  has_many :comments
+
   def publish
   end
 end